+++ src/Search/SimpleEntitySearchHelper.php
@@ -3,6 +3,7 @@
 namespace Becklyn\RadBundle\Search;
 
 use Doctrine\ORM\Query\Expr\Comparison;
+use Doctrine\ORM\Query\Expr\Func;
 use Doctrine\ORM\Query\Expr\Orx;
 use Doctrine\ORM\QueryBuilder;
 
@@ -20,7 +21,8 @@ final class SimpleEntitySearchHelper
 
 
     /**
-     * Applies the search to the given query builder
+     * Applies the search to the given query builder.
+     * Searches single text fields.
      */
     public function applySearch (QueryBuilder $queryBuilder, ?string $query, array $fields) : QueryBuilder
     {
@@ -38,6 +40,43 @@ final class SimpleEntitySearchHelper
         {
             $expr->add(
                 new Comparison($field, "LIKE", ":__searchTerm")
+            );
+        }
+
+        return $queryBuilder
+            ->andWhere($expr)
+            ->setParameter("__searchTerm", $tokenized);
+    }
+
+
+    /**
+     * Applies the search to the given query builder.
+     * Searches a JSON structure for strings matching the query.
+     */
+    public function applyJsonSearch (QueryBuilder $queryBuilder, ?string $query, array $fields) : QueryBuilder
+    {
+        $query = \trim((string) $query);
+
+        if ("" === $query)
+        {
+            return $queryBuilder;
+        }
+
+        $tokenized = $this->tokenizer->transformToQuery($query);
+        $expr = new Orx();
+
+        foreach ($fields as $field)
+        {
+            $expr->add(
+                new Comparison(
+                    new Func("JSON_SEARCH", [
+                        $field,
+                        "'one'",
+                        ":__searchTerm",
+                    ]),
+                    "IS NOT",
+                    "NULL"
+                )
             );
         }